Added integer size_t

+++ src/vsTTempSmooth_AVX2.cpp
@@ -545,7 +545,7 @@ float ComparePlane_avx2(PVideoFrame& src, PVideoFrame& src1, const int bits_per_
 
     Vec8f accum{ 0.0f };
 
-    for (size_t y{ 0 }; y < height; ++y)
+    for (size_t y{ 0 }; y < static_cast<size_t>(height); ++y)
     {
         for (size_t x{ 0 }; x < width; x += 8)
         {
